+++ .github/CHANGELOG.md
@@ -13,6 +13,7 @@
 - fix: #534
 - fix: #535 Application menu is not updated when switching windows
 - fix #216 and #311 key binding issues on Linux and Windows
+- fix #546 paste issue in table
 
 ### 0.13.50